+++ src/Server/Persistence/Systems/Persistence/PersistenceSystem.cs
@@ -86,6 +86,7 @@ public sealed class PersistenceSystem : ISystem
         EventId.Core_Quit,
         EventId.Server_Persistence_RetrieveEntity,
         EventId.Server_Persistence_RetrieveEntitiesInRange,
+        EventId.Server_Persistence_RetrieveWorldSegment,
         EventId.Server_Persistence_Synchronize,
         EventId.Server_Accounts_SelectPlayer
     };