+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 #### 14 January 2024
 
 * Fix issue in `PlayerComponentEventFilter` that caused event triggers for new player characters to be missed.
+* Fix various issues with world segment subscriptions and entity synchronization.
+* Begin fixing MessagePack issues with new event types. Still need to add support for serializing `Vector3<T>`.
 
 #### 13 January 2024